Issue 55: Extract all logic pertaining to error message generation from the rule.
The rule should not know anything about error messages. This is part of the responsibility of the module calculating rule violations. This module has all information that should be part of the error message and can simply generate it directly without the indirection via Rule.
