+++ src/pytestarch/query_language/rule.py
@@ -14,9 +14,6 @@ from pytestarch.query_language.base_language import (
     RuleSubject,
 )
 from pytestarch.query_language.exceptions import ImproperlyConfigured
-from pytestarch.rule_assessment.error_message.message_generator import (
-    RuleViolationMessageGenerator,
-)
 from pytestarch.rule_assessment.rule_check.behavior_requirement import (
     BehaviorRequirement,
 )
@@ -25,7 +22,6 @@ from pytestarch.rule_assessment.rule_check.rule_matcher import (
     DefaultRuleMatcher,
     RuleMatcher,
 )
-from pytestarch.rule_assessment.rule_check.rule_violations import RuleViolations
 
 
 @dataclass
@@ -159,10 +155,7 @@ class Rule(
         self._assert_required_configuration_present()
 
         matcher = self._prepare_rule_matcher()
-        rule_violations = matcher.match(evaluable)
-
-        if rule_violations:
-            raise AssertionError(self._create_rule_violation_message(rule_violations))
+        matcher.match(evaluable)
 
     def _prepare_rule_matcher(self) -> RuleMatcher:
         module_requirement = ModuleRequirement(
@@ -252,17 +245,6 @@ class Rule(
     @classmethod
     def _name_or_empty(cls, empty: bool, clz: type) -> str:
         return f"a {clz.__name__}" if empty else ""
-
-    def _create_rule_violation_message(self, rule_violations: RuleViolations) -> str:
-        message_generator = RuleViolationMessageGenerator(
-            self._configuration.module_to_check,
-            self._configuration.modules_to_check_against,
-            self._configuration.import_,
-        )
-        single_violation_messages = message_generator.create_rule_violation_messages(
-            rule_violations
-        )
-        return "\n".join(single_violation_messages)
 
     @classmethod
     def _convert_aliases(cls, configuration: RuleConfiguration) -> RuleConfiguration:
