+++ src/pytestarch/rule_assessment/rule_check/rule_matcher.py
@@ -6,6 +6,9 @@ from pytestarch.eval_structure.evaluable_architecture import (
     LaxDependenciesByBaseModule,
     StrictDependenciesByBaseModules,
 )
+from pytestarch.rule_assessment.error_message.message_generator import (
+    RuleViolationMessageGenerator,
+)
 from pytestarch.rule_assessment.rule_check.behavior_requirement import (
     BehaviorRequirement,
 )
@@ -25,28 +28,45 @@ class RuleMatcher(ABC):
         self._behavior_requirement = behavior_requirement
 
     @abstractmethod
-    def match(self, evaluable: EvaluableArchitecture) -> RuleViolations:
+    def match(self, evaluable: EvaluableArchitecture) -> None:
         """
         Checks whether an expected behavior is exhibited by the EvaluableArchitecture.
+        If there are any rule violations, will raise an error detailing the violations.
 
         Args:
             evaluable: object to check
-        Returns:
-            overview of rule violations
+        Raises:
+            AssertionError
         """
         raise NotImplementedError()
 
 
 class DefaultRuleMatcher(RuleMatcher):
-    def match(self, evaluable: EvaluableArchitecture) -> RuleViolations:
-        strict_dependencies = self._get_strict_dependencies(evaluable)
+    def match(self, evaluable: EvaluableArchitecture) -> None:
+        rule_violations = self._find_rule_violations(evaluable)
 
+        if rule_violations:
+            raise AssertionError(self._create_rule_violation_message(rule_violations))
+
+    def _find_rule_violations(self, evaluable: EvaluableArchitecture) -> RuleViolations:
+        strict_dependencies = self._get_strict_dependencies(evaluable)
         lax_dependencies = self._get_lax_dependencies(evaluable)
 
         return self._behavior_requirement.generate_rule_violation(
             strict_dependencies,
             lax_dependencies,
         )
+
+    def _create_rule_violation_message(self, rule_violations: RuleViolations) -> str:
+        message_generator = RuleViolationMessageGenerator(  # TODO
+            self._module_requirement._left_hand_modules_as_specified_by_user,
+            self._module_requirement._right_hand_modules_as_specified_by_user,
+            self._module_requirement._right_hand_module_has_specifier,
+        )
+        single_violation_messages = message_generator.create_rule_violation_messages(
+            rule_violations
+        )
+        return "\n".join(single_violation_messages)
 
     def _get_lax_dependencies(
         self, evaluable: EvaluableArchitecture
